+++ src/06-group/controller.ts
@@ -1,4 +1,5 @@
 import { IRequestWithAuth } from '../root/app.interfaces';
+import { BadRequestError } from '../root/responseHandler/error.response';
 import { CREATED, OK } from '../root/responseHandler/success.response';
 import { Fn, catchAsync } from '../root/utils/catchAsync';
 import { IGroupController } from './interfaces/controller';
@@ -6,9 +7,11 @@ import GroupService from './service';
 
 class GroupController<T extends IRequestWithAuth> implements IGroupController<T> {
   createOne: Fn<T> = catchAsync(async (req, res, next) => {
+    const { groups } = req.body;
+    if (!groups) throw new BadRequestError('Invalid transmitted data');
     const createdNewGroup = await GroupService.createGroup({
       boardId: req.params.boardId,
-      groups: req.body.groups,
+      groups,
     });
 
     new CREATED({
@@ -34,9 +37,11 @@ class GroupController<T extends IRequestWithAuth> implements IGroupController<T>
   });
 
   updateAllGroups: Fn<T> = catchAsync(async (req, res, next) => {
+    const { groups } = req.body;
+    if (!groups) throw new BadRequestError('Invalid transmitted data');
     const updatedAllGroups = await GroupService.updateAllGroups({
       boardId: req.params.boardId,
-      groups: req.body.groups,
+      groups,
     });
     new OK({
       message: 'Update all groups successfully',
@@ -47,10 +52,12 @@ class GroupController<T extends IRequestWithAuth> implements IGroupController<T>
   });
 
   deleteOne: Fn<T> = catchAsync(async (req, res, next) => {
+    const { groups } = req.body;
+    if (!groups) throw new BadRequestError('Invalid transmitted data');
     await GroupService.deleteGroup({
       boardId: req.params.boardId,
       groupId: req.params.id,
-      groups: req.body.groups,
+      groups,
     });
     new OK({
       message: 'Delete group successfully',
